feat: getHolderAndDuration moved out from nftSDK class
