+++ src/nftSDKFunctions/index.ts
@@ -17,7 +17,7 @@
  * limitations under the License.
  *
  */
-import { Client, NftId, PrivateKey } from '@hashgraph/sdk';
+import { Client, NftId } from '@hashgraph/sdk';
 import { CreateCollectionKeysType, CustomFeeType } from '../types/create-collection';
 import { Network } from '../types/mint-token';
 import { createCollectionFunction } from './create-collection';
@@ -30,8 +30,6 @@ import { estimateNftMintingInHbar } from './estimate-nft-minting-in-hbar';
 import { estimateNftMintingInDollars } from './estimate-nft-minting-in-dollars';
 import { estimateCreateCollectionInDollars } from './estimate-create-collection-in-dollars';
 import { estimateCreateCollectionInHbar } from './estimate-create-collection-in-hbar';
-import { getHolderAndDuration } from './get-holder-and-duration';
-import { NetworkName } from '@hashgraph/sdk/lib/client/Client';
 import { getPrivateKeyFromString } from '../helpers/get-private-key-from-string';
 
 export class HederaNFTSDK {
@@ -223,9 +221,5 @@ export class HederaNFTSDK {
       supplyKey: supplyKey ? getPrivateKeyFromString(supplyKey) : getPrivateKeyFromString(this.privateKey),
       mirrorNodeUrl: this.mirrorNodeUrl,
     });
-  }
-
-  getHolderAndDuration({ tokenId, serialNumber, network = 'mainnet' }: { tokenId: string; serialNumber: number; network?: NetworkName }) {
-    return getHolderAndDuration({ tokenId, serialNumber, network });
   }
 }
